requirements.txt
gitpython
rich
requests